+++ ConsoleApp1/Tools/Interface/IConfigurationParameters.cs
@@ -12,7 +12,7 @@ namespace Tools.Interface
         public string OutputFileNameDateFormat { get; set; }
         public string OutputFileNameTimeFormat { get; set; }
         public string OutputSavePath { get; set; }
-        public int TimeInterval { get; set; }
+        public decimal TimeInterval { get; set; }
         public string HeadersTitle { get; set; }
         public string LogFilePath { get; set; }
         public string LogfileName { get; set; }
